requirements_imageai_216.txt
imageai==2.1.6
h5py==2.10.0
keras==2.4.3
matplotlib==3.3.2
numpy==1.19.3
pillow==7.0.0
scipy==1.4.1
opencv-python>=4.1.2
